requirements.txt
Django==2.0.5
djangorestframework==3.8.2
Scrapy==1.5.0
newspaper3k==0.2.6
dill==0.2.8.2
nltk==3.3
scikit-learn==0.19.1
numpy==1.14.3
psycopg2>=2.7,<2.8 --no-binary psycopg2
Unidecode==1.0.22
dateparser==0.7.0:w
